Rename to uinstrument to avoid clashing with new haxelib
I never published this and someone has just recently pushing another
library using the `instrument` on haxelib.

While that library serves a similar purpose, I like the flexibility in
ours, and I think it still has usefulness as the *micro* instrumentation
library it always claimed to be.

Rename ours to uinstrument to avoid the name clash.